meld ok

+++ tRelatorio.h
@@ -25,10 +25,10 @@ int compararDiagnostico(const void *a, const void *b);
 tDiagnostico** CriaVetorDiagnostico();
 tDiagnostico* CriaDiagnostico(char* nome);
 tRelatorio* CriaRelatorio(tClinica* clinica);
-void calculaEstatisticasProcedimentos(tLesoes **vetor, int numLesoes,float *totalLesoes ,float *numCirurgias, float *numCrioterapias);
-void calculaEstatisticasTamanhoLesoes(tLesoes **vetor, int numLesoes, float *media, float *desvioPadrao);
-void calcularDistribuicaoGenero(tPaciente **pacientes, int numPacientes, float *masculino, float *feminino, float *outros);
-void calculaEstatisticasIdade(tPaciente **pacientes, int numPacientes, float *media, float *desvioPadrao);
+void calculaEstatisticasProcedimentos(tLesoes **vetor, int numLesoes,int *totalLesoes ,int *numCirurgias, int *numCrioterapias);
+void calculaEstatisticasTamanhoLesoes(tLesoes **vetor, int numLesoes, int *media, int *desvioPadrao);
+void calcularDistribuicaoGenero(tPaciente **pacientes, int numPacientes, int *masculino, int *feminino, int *outros);
+void calculaEstatisticasIdade(tPaciente **pacientes, int numPacientes, int *media, int *desvioPadrao);
 
 void desalocaRelatorio(void *dado);
 void imprimeNaTelaRelatorio(void *dado);
